@@ -1,8 +1,8 @@
 {
     "name": "park-manager/hub-kit",
     "description": "HubKit allows project(s) maintainers to easily manage their GitHub repositories",
-    "type": "project",
     "license": "MIT",
+    "type": "project",
     "authors": [
         {
             "name": "Sebastiaan Stok",
@@ -10,26 +10,38 @@
         }
     ],
     "require": {
-        "php": "^7.2|^8.0",
-        "symfony/process": "~3.4",
-        "symfony/event-dispatcher": "~3.4",
-        "symfony/filesystem": "~3.4",
-        "symfony/console": "^3.4",
-        "webmozart/console": "dev-master@dev",
-        "pimple/pimple": "^3.0",
-        "guzzlehttp/psr7": "^1.2",
+        "php": "^8.1",
+        "composer/semver": "^3.3.2",
+        "ddd/slug": "~1.0.0",
         "guzzlehttp/guzzle": "^7.0.1",
+        "guzzlehttp/psr7": "^1.2",
         "http-interop/http-factory-guzzle": "^1.0",
         "knplabs/github-api": "^3.0",
-        "ddd/slug": "~1.0.0",
+        "pimple/pimple": "^3.0",
         "rollerworks/version": "^0.3.1",
-        "composer/semver": "^1.5"
+        "symfony/console": "^6.0",
+        "symfony/event-dispatcher": "^6.0",
+        "symfony/filesystem": "^6.0",
+        "symfony/process": "^6.0",
+        "webmozart/console": "^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.3.1|^8.5.12",
-        "symfony/phpunit-bridge": "^4.1",
-        "symfony/debug": "^3.4"
+        "jangregor/phpstan-prophecy": "^1.0.0",
+        "phpspec/prophecy-phpunit": "^v2.0.1",
+        "phpstan/phpstan": "^1.8.1",
+        "phpstan/phpstan-phpunit": "^1.1.1",
+        "phpunit/phpunit": "^9.5.21",
+        "symfony/error-handler": "^6.0",
+        "symfony/phpunit-bridge": "^6.0"
     },
+    "repositories": [
+        {
+            "type": "github",
+            "url": "https://github.com/park-manager/webmozart-console.git"
+        }
+    ],
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "autoload": {
         "psr-4": {
             "HubKit\\": "src"
@@ -40,12 +52,15 @@
             "HubKit\\Tests\\": "tests"
         }
     },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
+    "config": {
+        "allow-plugins": {
+            "symfony/flex": true
+        },
+        "sort-packages": true
+    },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0-dev",
-            "dev-1.x": "1.1-dev"
+            "dev-1.x": "1.2-dev"
         }
     }
 }
