@@ -43,24 +43,24 @@ final class SwitchBaseHandler extends GitBaseHandler
 
         $branch = $pullRequest['head']['ref'];
         $remote = $pullRequest['head']['user']['login'];
-        $tmpBranch = '_temp_'.$remote.'--'.$branch.'--'.$newBase;
+        $tmpBranch = '_temp_' . $remote . '--' . $branch . '--' . $newBase;
 
         $this->git->ensureRemoteExists($remote, $pullRequest['head']['repo']['ssh_url']);
         $this->git->remoteUpdate($remote);
         $this->git->remoteUpdate('upstream');
 
         // Operation was already in progress (but gave a conflict).
-        if (file_exists($this->git->getGitDirectory().'/.hubkit-switch')) {
+        if (file_exists($this->git->getGitDirectory() . '/.hubkit-switch')) {
             $this->handleIncompleteOperation($remote, $tmpBranch, $branch);
         }
 
-        $this->switchBranchBase($remote, $branch, 'upstream/'.$pullRequest['base']['ref'], $newBase, $tmpBranch);
+        $this->switchBranchBase($remote, $branch, 'upstream/' . $pullRequest['base']['ref'], $newBase, $tmpBranch);
         $this->pushToRemote($remote, $tmpBranch, $branch);
         $this->deleteTempBranch($tmpBranch);
 
         $this->github->updatePullRequest($pullRequest['number'], ['base' => $newBase]);
 
-        if (!$args->getOption('skip-help')) {
+        if (! $args->getOption('skip-help')) {
             $this->postHelpComment($pullRequest, $branch);
         }
 
@@ -83,7 +83,7 @@ final class SwitchBaseHandler extends GitBaseHandler
     private function handleIncompleteOperation(string $remote, string $tmpBranch, string $branch): void
     {
         $gitDir = $this->git->getGitDirectory();
-        $tmpWorkingBranch = trim(file_get_contents($gitDir.'/.hubkit-switch'));
+        $tmpWorkingBranch = trim(file_get_contents($gitDir . '/.hubkit-switch'));
 
         // If branch was switched, assume the operation is aborted completely (and start over from scratch).
         // If branch name equals, assume a resolve - check if there are actual changes (before we push with force)
@@ -100,7 +100,7 @@ final class SwitchBaseHandler extends GitBaseHandler
                 return; // Delete always performed in switchBranchBase()
             }
 
-            if (!$this->style->confirm('Do you want to continue the previous operation?')) {
+            if (! $this->style->confirm('Do you want to continue the previous operation?')) {
                 throw new \RuntimeException(
                     'Failed! Cannot perform switch while another operation is still pending. Please abort previous operation first.'
                 );
@@ -120,33 +120,33 @@ final class SwitchBaseHandler extends GitBaseHandler
     {
         $activeBranch = $this->git->getActiveBranchName();
 
-        if ('_' === $activeBranch[0]) {
+        if ($activeBranch[0] === '_') {
             $activeBranch = 'master';
         }
 
         // Always (re)start the rebase process from scratch in case something went horrible wrong.
         $this->deleteTempBranch($tmpBranch);
-        $this->git->checkout($remote.'/'.$branch);
+        $this->git->checkout($remote . '/' . $branch);
         $this->git->checkout($tmpBranch, true);
 
-        file_put_contents($this->git->getGitDirectory().'/.hubkit-switch', $tmpBranch);
+        file_put_contents($this->git->getGitDirectory() . '/.hubkit-switch', $tmpBranch);
 
         try {
-            $this->process->mustRun(['git', 'rebase', '--onto', 'upstream/'.$newBase, $currentBase, $tmpBranch]);
+            $this->process->mustRun(['git', 'rebase', '--onto', 'upstream/' . $newBase, $currentBase, $tmpBranch]);
             $this->git->checkout($activeBranch);
         } catch (ProcessFailedException $e) {
             throw new \RuntimeException(
                 <<<MESSAGE
-Git rebase operation failed: {$e->getMessage()}
+                    Git rebase operation failed: {$e->getMessage()}
 
--------------------------------------------------------------
-Please resolve the conflicts manually, `git add` the resolved
-files AND execute `git rebase --continue` afterwards.
+                    -------------------------------------------------------------
+                    Please resolve the conflicts manually, `git add` the resolved
+                    files AND execute `git rebase --continue` afterwards.
 
-Then run the operation again.
+                    Then run the operation again.
 
-DO NOT PUSH THE CHANGES MANUALLY!
-MESSAGE
+                    DO NOT PUSH THE CHANGES MANUALLY!
+                    MESSAGE
                 , 0, $e
             );
         }
@@ -154,10 +154,10 @@ MESSAGE
 
     private function guardWorkingTreeIsReady(): void
     {
-        if (!$this->git->isWorkingTreeReady()) {
+        if (! $this->git->isWorkingTreeReady()) {
             throw new \RuntimeException(
-                'The Git working tree is not ready. There are uncommitted changes or a rebase is in progress.'.
-                "\n".
+                'The Git working tree is not ready. There are uncommitted changes or a rebase is in progress.' .
+                "\n" .
                 'If there were conflicts during the switch run `git rebase --continue` and run the `switch-base` command again.'
             );
         }
@@ -166,17 +166,17 @@ MESSAGE
     private function deleteTempBranch(string $tmpBranch): void
     {
         $this->process->run(['git', 'branch', '-D', $tmpBranch]);
-        @unlink($this->git->getGitDirectory().'/.hubkit-switch');
+        @unlink($this->git->getGitDirectory() . '/.hubkit-switch');
     }
 
     private function pushToRemote(string $remote, string $tmpBranch, string $branch): void
     {
-        $this->process->mustRun(['git', 'push', '--force', $remote, $tmpBranch.':'.$branch], 'Push failed (access disabled?)');
+        $this->process->mustRun(['git', 'push', '--force', $remote, $tmpBranch . ':' . $branch], 'Push failed (access disabled?)');
     }
 
     private function guardNotClosed(string $state): void
     {
-        if ('open' !== $state) {
+        if ($state !== 'open') {
             throw new \InvalidArgumentException('Cannot switch base of closed/merged pull-request.');
         }
     }
@@ -187,7 +187,7 @@ MESSAGE
             throw new \InvalidArgumentException(sprintf('Cannot switch base, current base is already "%s".', $newBase));
         }
 
-        if (!$this->git->remoteBranchExists('upstream', $newBase)) {
+        if (! $this->git->remoteBranchExists('upstream', $newBase)) {
             throw new \InvalidArgumentException(sprintf('Cannot switch base, base branch "%s" does not exists.', $newBase));
         }
     }
@@ -201,15 +201,15 @@ MESSAGE
         $this->github->createComment(
             $pullRequest['number'],
             <<<MESSAGE
-The base of this pull-request was changed, you need fetch and reset your local branch 
-if you want to add new commits to this pull request. **Reset before you pull, else commits
-may become messed-up.**
+                The base of this pull-request was changed, you need fetch and reset your local branch
+                if you want to add new commits to this pull request. **Reset before you pull, else commits
+                may become messed-up.**
 
-Unless you added new commits (to this branch) locally that you did not push yet, 
-execute `git fetch origin && git reset "{$branch}"` to update your local branch.
- 
-Feel free to ask for assistance when you get stuck :+1:
-MESSAGE
+                Unless you added new commits (to this branch) locally that you did not push yet,
+                execute `git fetch origin && git reset "{$branch}"` to update your local branch.
+
+                Feel free to ask for assistance when you get stuck :+1:
+                MESSAGE
         );
     }
 }
