@@ -41,14 +41,25 @@ class StatusTable
         $this->output = $output;
     }
 
-    public function render()
+    public function render(): void
     {
         $table = new Table($this->output);
         $table->getStyle()
-            ->setHorizontalBorderChar('-')
-            ->setVerticalBorderChar('')
-            ->setCrossingChar('')
-            ->setCellRowContentFormat('  %s  ');
+            ->setCrossingChars(
+                cross: '',
+                topLeft: '',
+                topMid: '',
+                topRight: '',
+                midRight: '',
+                bottomRight: '',
+                bottomMid: '',
+                bottomLeft: '',
+                midLeft: '',
+            )
+            ->setHorizontalBorderChars('-', '-')
+            ->setVerticalBorderChars(outside: '', inside: '')
+            ->setCellRowContentFormat('  %s  ')
+        ;
 
         $table->setHeaders(['Item', 'Status', 'Details']);
         $table->setRows($this->rows);
@@ -63,7 +74,7 @@ class StatusTable
     }
 
     // make other methods none static
-    public function addRow(string $label, string $status, ?string $message = null)
+    public function addRow(string $label, string $status, ?string $message = null): void
     {
         $this->rows[] = [$label, self::STATUS_LABELS[$status], wordwrap((string) $message, 38)];
         $this->statuses[$status] = true;
