@@ -19,7 +19,10 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Style\StyleInterface;
 use Symfony\Component\Process\Exception\ProcessFailedException;
 
-class GitBranchTest extends TestCase
+/**
+ * @internal
+ */
+final class GitBranchTest extends TestCase
 {
     use GitTesterTrait;
 
@@ -33,11 +36,11 @@ class GitBranchTest extends TestCase
     /** @before */
     public function setUpLocalRepository(): void
     {
-        $this->cwd = $this->localRepository = $this->createGitDirectory($this->getTempDir().'/git');
+        $this->cwd = $this->localRepository = $this->createGitDirectory($this->getTempDir() . '/git');
         $this->commitFileToRepository('foo.txt', $this->localRepository);
         $this->commitFileToRepository('diggy.txt', $this->localRepository);
 
-        $this->remoteRepository = $this->createBareGitDirectory($this->getTempDir().'/git2');
+        $this->remoteRepository = $this->createBareGitDirectory($this->getTempDir() . '/git2');
         $this->addRemote('origin', $this->remoteRepository, $this->localRepository);
         $this->runCliCommand(['git', 'push', 'origin', 'master'], $this->localRepository);
 
@@ -45,14 +48,14 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_returns_up_to_date_when_up_to_date()
+    public function it_returns_up_to_date_when_up_to_date(): void
     {
         self::assertEquals(GitBranch::STATUS_UP_TO_DATE, $this->git->getRemoteDiffStatus('origin', 'master'));
         self::assertEquals(GitBranch::STATUS_UP_TO_DATE, $this->git->getRemoteDiffStatus('origin', 'master', 'master'));
     }
 
     /** @test */
-    public function it_returns_needs_push_when_local_is_ahead()
+    public function it_returns_needs_push_when_local_is_ahead(): void
     {
         $this->commitFileToRepository('hole.cmd', $this->localRepository);
 
@@ -60,7 +63,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_returns_needs_pull_when_remote_is_ahead()
+    public function it_returns_needs_pull_when_remote_is_ahead(): void
     {
         $this->runCliCommand(['git', 'reset', '--hard', 'HEAD@{1}'], $this->localRepository);
 
@@ -68,7 +71,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_returns_needs_divered_when_both_are_newer()
+    public function it_returns_needs_divered_when_both_are_newer(): void
     {
         $this->runCliCommand(['git', 'reset', '--hard', 'HEAD@{1}'], $this->localRepository);
         $this->commitFileToRepository('something.txt', $this->localRepository);
@@ -77,19 +80,19 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_gets_the_active_branch_name()
+    public function it_gets_the_active_branch_name(): void
     {
         self::assertEquals('master', $this->git->getActiveBranchName());
     }
 
     /** @test */
-    public function it_cannot_get_active_branch_when_in_deteached_head()
+    public function it_cannot_get_active_branch_when_in_deteached_head(): void
     {
         $this->runCliCommand(['git', 'checkout', 'HEAD@{1}']);
 
         $this->expectException('RuntimeException');
         $this->expectExceptionMessage(
-            'You are currently in a detached HEAD state, '.
+            'You are currently in a detached HEAD state, ' .
             'unable to get active branch-name.Please run `git checkout` first.'
         );
 
@@ -97,7 +100,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_cannot_get_latest_tag_if_none_exists()
+    public function it_cannot_get_latest_tag_if_none_exists(): void
     {
         $this->expectException(ProcessFailedException::class);
         $this->expectExceptionMessage('fatal: No names found, cannot describe anything.');
@@ -106,7 +109,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_gets_latest_tag()
+    public function it_gets_latest_tag(): void
     {
         $this->runCliCommand(['git', 'tag', 'v1.0']);
         $this->commitFileToRepository('you', $this->localRepository);
@@ -116,7 +119,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_gets_versioned_branches_in_correct_order()
+    public function it_gets_versioned_branches_in_correct_order(): void
     {
         $this->addRemote('upstream', $this->remoteRepository);
         $this->givenRemoteBranchesExist(['1.0', 'v1.1', '2.0', 'x.1']);
@@ -126,7 +129,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_returns_whether_remote_branch_exists()
+    public function it_returns_whether_remote_branch_exists(): void
     {
         $this->setUpstreamRepository();
         $this->givenRemoteBranchesExist(['2.0', '1.x']);
@@ -143,7 +146,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_returns_whether_local_branch_exists()
+    public function it_returns_whether_local_branch_exists(): void
     {
         $this->givenLocalBranchesExist(['2.0', '1.x']);
 
@@ -154,7 +157,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_removes_remote_branch()
+    public function it_removes_remote_branch(): void
     {
         $this->givenRemoteBranchesExist(['2.0', '1.x']);
 
@@ -166,7 +169,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_removes_local_branch()
+    public function it_removes_local_branch(): void
     {
         $this->givenLocalBranchesExist(['2.0', '1.x']);
 
@@ -178,7 +181,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_removes_local_branch_respecting_the_merge_status()
+    public function it_removes_local_branch_respecting_the_merge_status(): void
     {
         $this->givenLocalBranchesExist(['2.0', '1.x']);
         $this->runCliCommand(['git', 'checkout', '2.0']);
@@ -192,7 +195,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_checkouts_out_an_existing_branch()
+    public function it_checkouts_out_an_existing_branch(): void
     {
         $this->givenLocalBranchesExist(['2.0']);
 
@@ -202,7 +205,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_checkouts_out_a_new_branch()
+    public function it_checkouts_out_a_new_branch(): void
     {
         $this->runCliCommand(['git', 'checkout', 'HEAD@{1}']);
 
@@ -212,7 +215,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_checkouts_out_a_remote_branche_existing_locally()
+    public function it_checkouts_out_a_remote_branche_existing_locally(): void
     {
         $this->givenRemoteBranchesExist(['2.0']);
         $this->givenLocalBranchesExist(['2.0']);
@@ -223,7 +226,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_checkouts_out_a_remote_branche()
+    public function it_checkouts_out_a_remote_branche(): void
     {
         $this->givenRemoteBranchesExist(['2.0']);
 
