@@ -16,28 +16,34 @@ namespace HubKit\Tests\Handler;
 use HubKit\Cli\Handler\BranchAliasHandler;
 use HubKit\Service\Git;
 use PHPUnit\Framework\TestCase;
+use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Webmozart\Console\Api\Args\Args;
 use Webmozart\Console\Api\Args\Format\ArgsFormat;
 use Webmozart\Console\Api\Args\Format\Argument;
 use Webmozart\Console\IO\BufferedIO;
 
-class BranchAliasHandlerTest extends TestCase
+/**
+ * @internal
+ */
+final class BranchAliasHandlerTest extends TestCase
 {
+    use ProphecyTrait;
+
     /** @var ObjectProphecy */
     private $git;
     /** @var BufferedIO */
     private $io;
 
     /** @before */
-    public function setUpCommandHandler()
+    public function setUpCommandHandler(): void
     {
         $this->git = $this->prophesize(Git::class);
         $this->io = new BufferedIO();
     }
 
     /** @test */
-    public function it_sets_a_new_alias()
+    public function it_sets_a_new_alias(): void
     {
         $this->git->getGitConfig('branch.master.alias')->willReturn('1.0-dev');
         $this->git->setGitConfig('branch.master.alias', '1.2-dev', true)->shouldBeCalled();
@@ -47,7 +53,7 @@ class BranchAliasHandlerTest extends TestCase
     }
 
     /** @test */
-    public function it_gets_current_alias()
+    public function it_gets_current_alias(): void
     {
         $this->git->getGitConfig('branch.master.alias')->willReturn('1.5-dev');
         $this->executeHandler();
@@ -56,7 +62,7 @@ class BranchAliasHandlerTest extends TestCase
     }
 
     /** @test */
-    public function it_requires_alias_in_a_specific_format()
+    public function it_requires_alias_in_a_specific_format(): void
     {
         $this->git->getGitConfig('branch.master.alias')->willReturn('1.0-dev');
 
@@ -66,7 +72,7 @@ class BranchAliasHandlerTest extends TestCase
         $this->executeHandler('v1.2');
     }
 
-    private function executeHandler(string $alias = null)
+    private function executeHandler(string $alias = null): void
     {
         $format = ArgsFormat::build()
             ->addArgument(new Argument('alias', Argument::OPTIONAL | Argument::STRING))
