@@ -17,10 +17,16 @@ use HubKit\Helper\ChangelogRenderer;
 use HubKit\Service\Git;
 use HubKit\Service\GitHub;
 use PHPUnit\Framework\TestCase;
+use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 
+/**
+ * @internal
+ */
 final class ChangelogRendererTest extends TestCase
 {
+    use ProphecyTrait;
+
     private const COMMITS = [
         [
             'sha' => '2bee3b83a9b0073497f37acd4f0920ef61945552',
@@ -30,7 +36,7 @@ final class ChangelogRendererTest extends TestCase
 
                 Discussion
                 ----------
-            
+
                 |Q            |A  |
                 |---          |---|
                 |Bug Fix?     |no |
@@ -39,10 +45,10 @@ final class ChangelogRendererTest extends TestCase
                 |Deprecations?|yes|
                 |Fixed Tickets|   |
                 |License      |MIT|
-            
+
                 Commits
                 -------
-            
+
                 1b04532c8a09d9084abce36f8d9daf675f89eacc Introduce a new API for ValuesBag',
         ],
         [
@@ -93,10 +99,10 @@ labels: removed-deprecation,bc-break
     private $github;
 
     /** @before */
-    public function setUpCommandHandler()
+    public function setUpCommandHandler(): void
     {
         $this->git = $this->prophesize(Git::class);
-        $this->git->guardWorkingTreeReady()->willReturn(null);
+        $this->git->guardWorkingTreeReady()->will(static function (): void {});
 
         $this->github = $this->prophesize(GitHub::class);
         $this->github->getHostname()->willReturn('github.com');
@@ -105,69 +111,69 @@ labels: removed-deprecation,bc-break
     }
 
     /** @test */
-    public function it_renders_a_changelog_with_categories_excluding_empty()
+    public function it_renders_a_changelog_with_categories_excluding_empty(): void
     {
         $this->git->getLogBetweenCommits('base', 'head')->willReturn(self::COMMITS);
 
         self::assertEquals(<<<'LOG'
-### Security
-- Fix password-reset procedure ([sstok](https://github.com/sstok)) [#57](https://github.com/park-manager/hubkit/issues/57)
+            ### Security
+            - Fix password-reset procedure ([sstok](https://github.com/sstok)) [#57](https://github.com/park-manager/hubkit/issues/57)
 
-### Added
-- Introduce a new API for ValuesBag ([sstok](https://github.com/sstok)) [#93](https://github.com/park-manager/hubkit/issues/93)
+            ### Added
+            - Introduce a new API for ValuesBag ([sstok](https://github.com/sstok)) [#93](https://github.com/park-manager/hubkit/issues/93)
 
-### Changed
-- Clean up ([sstok2](https://github.com/sstok2)) [#56](https://github.com/park-manager/hubkit/issues/56)
+            ### Changed
+            - Clean up ([sstok2](https://github.com/sstok2)) [#56](https://github.com/park-manager/hubkit/issues/56)
 
-### Deprecated
-- Great new architecture ([sstok](https://github.com/sstok), [someone](https://github.com/someone)) [#55](https://github.com/park-manager/hubkit/issues/55)
+            ### Deprecated
+            - Great new architecture ([sstok](https://github.com/sstok), [someone](https://github.com/someone)) [#55](https://github.com/park-manager/hubkit/issues/55)
 
-### Removed
-- [BC BREAK] Removed deprecated API ([sstok](https://github.com/sstok)) [#52](https://github.com/park-manager/hubkit/issues/52)
-LOG
-, $this->getRenderer()->renderChangelogByCategories('base', 'head'));
+            ### Removed
+            - [BC BREAK] Removed deprecated API ([sstok](https://github.com/sstok)) [#52](https://github.com/park-manager/hubkit/issues/52)
+            LOG
+            , $this->getRenderer()->renderChangelogByCategories('base', 'head'));
     }
 
     /** @test */
-    public function it_renders_a_changelog_with_categories_including_empty()
+    public function it_renders_a_changelog_with_categories_including_empty(): void
     {
         $this->git->getLogBetweenCommits('base', 'head')->willReturn(self::COMMITS);
 
         self::assertEquals(<<<'LOG'
-### Security
-- Fix password-reset procedure ([sstok](https://github.com/sstok)) [#57](https://github.com/park-manager/hubkit/issues/57)
+            ### Security
+            - Fix password-reset procedure ([sstok](https://github.com/sstok)) [#57](https://github.com/park-manager/hubkit/issues/57)
 
-### Added
-- Introduce a new API for ValuesBag ([sstok](https://github.com/sstok)) [#93](https://github.com/park-manager/hubkit/issues/93)
+            ### Added
+            - Introduce a new API for ValuesBag ([sstok](https://github.com/sstok)) [#93](https://github.com/park-manager/hubkit/issues/93)
 
-### Changed
-- Clean up ([sstok2](https://github.com/sstok2)) [#56](https://github.com/park-manager/hubkit/issues/56)
+            ### Changed
+            - Clean up ([sstok2](https://github.com/sstok2)) [#56](https://github.com/park-manager/hubkit/issues/56)
 
-### Deprecated
-- Great new architecture ([sstok](https://github.com/sstok), [someone](https://github.com/someone)) [#55](https://github.com/park-manager/hubkit/issues/55)
+            ### Deprecated
+            - Great new architecture ([sstok](https://github.com/sstok), [someone](https://github.com/someone)) [#55](https://github.com/park-manager/hubkit/issues/55)
 
-### Removed
-- [BC BREAK] Removed deprecated API ([sstok](https://github.com/sstok)) [#52](https://github.com/park-manager/hubkit/issues/52)
+            ### Removed
+            - [BC BREAK] Removed deprecated API ([sstok](https://github.com/sstok)) [#52](https://github.com/park-manager/hubkit/issues/52)
 
-### Fixed
-- nothing
-LOG
-, $this->getRenderer()->renderChangelogByCategories('base', 'head', false));
+            ### Fixed
+            - nothing
+            LOG
+            , $this->getRenderer()->renderChangelogByCategories('base', 'head', false));
     }
 
     /** @test */
-    public function it_renders_a_changelog_without_categories()
+    public function it_renders_a_changelog_without_categories(): void
     {
         $this->git->getLogBetweenCommits('base', 'head')->willReturn(self::COMMITS);
 
         self::assertEquals(<<<'LOG'
-- Introduce a new API for ValuesBag ([sstok](https://github.com/sstok)) [#93](https://github.com/park-manager/hubkit/issues/93)
-- Clean up ([sstok2](https://github.com/sstok2)) [#56](https://github.com/park-manager/hubkit/issues/56)
-- Great new architecture ([sstok](https://github.com/sstok), [someone](https://github.com/someone)) [#55](https://github.com/park-manager/hubkit/issues/55)
-- [BC BREAK] Removed deprecated API ([sstok](https://github.com/sstok)) [#52](https://github.com/park-manager/hubkit/issues/52)
-- Fix password-reset procedure ([sstok](https://github.com/sstok)) [#57](https://github.com/park-manager/hubkit/issues/57)
-LOG
-, $this->getRenderer()->renderChangelogOneLine('base', 'head'));
+            - Introduce a new API for ValuesBag ([sstok](https://github.com/sstok)) [#93](https://github.com/park-manager/hubkit/issues/93)
+            - Clean up ([sstok2](https://github.com/sstok2)) [#56](https://github.com/park-manager/hubkit/issues/56)
+            - Great new architecture ([sstok](https://github.com/sstok), [someone](https://github.com/someone)) [#55](https://github.com/park-manager/hubkit/issues/55)
+            - [BC BREAK] Removed deprecated API ([sstok](https://github.com/sstok)) [#52](https://github.com/park-manager/hubkit/issues/52)
+            - Fix password-reset procedure ([sstok](https://github.com/sstok)) [#57](https://github.com/park-manager/hubkit/issues/57)
+            LOG
+            , $this->getRenderer()->renderChangelogOneLine('base', 'head'));
     }
 
     private function getRenderer(): ChangelogRenderer
