@@ -16,12 +16,18 @@ namespace HubKit\Tests\Unit\Service\Git;
 use HubKit\Service\CliProcess;
 use HubKit\Service\Git\GitBranch;
 use PHPUnit\Framework\TestCase;
+use Prophecy\PhpUnit\ProphecyTrait;
 use RuntimeException;
 use Symfony\Component\Console\Style\StyleInterface;
 use Symfony\Component\Process\Process;
 
-class GitBranchTest extends TestCase
+/**
+ * @internal
+ */
+final class GitBranchTest extends TestCase
 {
+    use ProphecyTrait;
+
     public function provideExpectedVersions(): array
     {
         return [
@@ -40,7 +46,7 @@ class GitBranchTest extends TestCase
      * @test
      * @dataProvider provideExpectedVersions
      */
-    public function it_gets_versioned_branches_in_correct_order(array $branches, array $expectedVersions)
+    public function it_gets_versioned_branches_in_correct_order(array $branches, array $expectedVersions): void
     {
         self::assertSame($expectedVersions, $this->createGitService($branches)->getVersionBranches('upstream'));
     }
@@ -53,13 +59,14 @@ class GitBranchTest extends TestCase
         $processHelper = $this->prophesize(CliProcess::class);
         $processHelper
             ->mustRun(['git', 'for-each-ref', '--format', '%(refname:strip=3)', 'refs/remotes/upstream'])
-            ->willReturn($process->reveal());
+            ->willReturn($process->reveal())
+        ;
 
         return new GitBranch($processHelper->reveal(), $this->createMock(StyleInterface::class));
     }
 
     /** @test */
-    public function it_syncs_nothing_when_diff_status_gives_up_to_date()
+    public function it_syncs_nothing_when_diff_status_gives_up_to_date(): void
     {
         $git = $this->givenGitRemoteDiffStatus(GitBranch::STATUS_UP_TO_DATE);
 
@@ -71,7 +78,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_syncs_pull_when_diff_status_gives_needs_pull()
+    public function it_syncs_pull_when_diff_status_gives_needs_pull(): void
     {
         $git = $this->givenGitRemoteDiffStatus(GitBranch::STATUS_NEED_PULL);
 
@@ -83,7 +90,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_syncs_push_when_diff_status_gives_needs_push_and_push_is_allowed()
+    public function it_syncs_push_when_diff_status_gives_needs_push_and_push_is_allowed(): void
     {
         $git = $this->givenGitRemoteDiffStatus(GitBranch::STATUS_NEED_PUSH);
 
@@ -95,7 +102,7 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_syncs_throws_when_diff_status_gives_needs_push_but_push_is_forbidden()
+    public function it_syncs_throws_when_diff_status_gives_needs_push_but_push_is_forbidden(): void
     {
         $git = $this->givenGitRemoteDiffStatus(GitBranch::STATUS_NEED_PUSH);
 
@@ -108,14 +115,14 @@ class GitBranchTest extends TestCase
     }
 
     /** @test */
-    public function it_syncs_throws_when_diff_status_gives_diverged()
+    public function it_syncs_throws_when_diff_status_gives_diverged(): void
     {
         $git = $this->givenGitRemoteDiffStatus(GitBranch::STATUS_DIVERGED);
 
         $this->expectException(RuntimeException::class);
         $this->expectExceptionMessage(
-            'Cannot safely perform the operation. '.
-            'Your local and remote version of branch "master" have differed.'.
+            'Cannot safely perform the operation. ' .
+            'Your local and remote version of branch "master" have differed.' .
             ' Please resolve this problem manually.'
         );
 
